Changes to long long prototypes.
git-svn-id: svn+ssh://svn.openmicroscopy.org.uk/home/svn/ome/trunk@2640 61f9eca8-113b-463c-a1a4-bab9b0a29f14

+++ src/C/omeis/omeis.h
@@ -34,7 +34,8 @@
  *------------------------------------------------------------------------------
  */
 
-typedef unsigned long long OID;
+/* typedef unsigned long long OID; */
+typedef u_int64_t OID;
 #define OME_IS_PIXL_SIG 0x5049584C /* PIXL in ASCII */
 #define OME_IS_PIXL_VER 1  /* Version 1 */
